reduce the path lengths ( trying to workaround #949 )
